// frontend/src/app/services/models/user-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UserDto {
  accountLocked?: boolean;
  address: string;
  cin: number;
  createdDate?: string;
  dateOfBirth: string;
  email: string;
  enable?: boolean;
  fullName: string;
  gender?: 'MALE' | 'FEMALE';
  id?: number;
  image?: string;
  lastModifiedDate?: string;
  phoneNumber: number;
  uid?: string;
}